% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vant.R
\name{alpha_vant}
\alias{alpha_vant}
\title{Alpha Vantage Function}
\usage{
alpha_vant(stock_symbol, frequency, interval = NULL)
}
\arguments{
\item{stock_symbol}{this is the ticker symbol for the object}

\item{frequency}{this refers to frequency of the data. Valid values include \code{intraday}, \code{daily}, \code{daily_adjusted}, \code{weekly}, \code{weekly_adjusted}, \code{monthly}, \code{monthly_adjusted},}

\item{interval}{in the case that the user desires intraday data, you may select the time intervals for the intraday data frame. Valid values include \code{1min}, \code{5min}, \code{15min}, \code{30min}, \code{60min}}
}
\value{
returns a data frame given the parameters of the API query.
}
\description{
Alpha Vantage Function
}
\examples{
alpha_vant("GOOGL", frequency = "daily") # a simple example

matrix1 = matrix(
  c(
    rep(1,3),
    rep(1,3),
    rep(1,3),
    rep(2,3),
    rep(2,3)
  ), 
  nrow = 5, ncol = 3, byrow = TRUE
)
with(subset(GOOGL_DAILY, GOOGL_DAILY$date > as.Date("2010-01-01")),
     {
       layout(matrix1)
       plot(date, open, type = "l", main = "GOOGL")
       lines(date, high, col = "blue", lwd = 0.3)
       lines(date, low, col = "blue", lwd = 0.3)
       
       plot(date, volume, type = "l", main = "GOOGL", col = "#108a4c")
     }
)


}
