% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_vant_auth.R
\name{alpha_vantage_auth}
\alias{alpha_vantage_auth}
\title{Initialize API Query}
\usage{
alpha_vantage_auth(x)
}
\arguments{
\item{x}{}
}
\value{
sets an environmental variable that safely stores your API key. Should not appear in your environment.
}
\description{
Initialize API Query
}
